+++ sandbox_town_backend/src/main/java/com/shuidun/sandbox_town_backend/bean/BuildingTypeDo.java
@@ -22,9 +22,9 @@ public class BuildingTypeDo {
 
     private String imagePath;
 
-    private Integer basicWidth;
+    private Double basicWidth;
 
-    private Integer basicHeight;
+    private Double basicHeight;
 
     private Integer rarity;
 }
