+++ sandbox_town_backend/src/main/java/com/shuidun/sandbox_town_backend/bean/MoveDto.java
@@ -9,13 +9,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class MoveDto {
     // 出发点x坐标
-    private Integer x0;
+    private Double x0;
     // 出发点y坐标
-    private Integer y0;
+    private Double y0;
     // 终点x坐标
-    private Integer x1;
+    private Double x1;
     // 终点y坐标
-    private Integer y1;
+    private Double y1;
     // 终点建筑ID（如果有）
     private String destBuildingId;
     // 终点精灵ID（如果有）
