+++ sandbox_town_backend/src/main/java/com/shuidun/sandbox_town_backend/bean/SpriteTypeDo.java
@@ -38,9 +38,9 @@ public class SpriteTypeDo {
 
     private Integer basicSpeed;
 
-    private Integer basicWidth;
+    private Double basicWidth;
 
-    private Integer basicHeight;
+    private Double basicHeight;
 
     private Integer basicVisionRange;
